backend/requirements.txt
fastapi>=0.104.0
sqlmodel>=0.0.8
uvicorn[standard]>=0.24.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
cryptography==41.0.7
PyJWT==2.8.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
email-validator>=2.0.0
